% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_plot_modules.R
\name{compute_module_scores}
\alias{compute_module_scores}
\title{Compute module scores and finalize module order}
\usage{
compute_module_scores(
  expr,
  gene_list,
  module_order = c("Proliferating-like", "Osteoblast-like", "Chondroblast-like",
    "Fibroblast-like")
)
}
\arguments{
\item{expr}{matrix (genes x samples), assumed already log1p + row-wise z-scored}

\item{gene_list}{named list of gene vectors}

\item{module_order}{optional desired order of modules}
}
\value{
list(scores=data.frame, modules_final=character)
}
\description{
Compute module scores and finalize module order
}
