% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_plot_modules.R
\name{order_and_prepare_heatmap}
\alias{order_and_prepare_heatmap}
\title{Order samples and assemble heatmap inputs}
\usage{
order_and_prepare_heatmap(scores, modules_final)
}
\arguments{
\item{scores}{data.frame after add_simplicity_scores()}

\item{modules_final}{character vector of module names}
}
\value{
list(mat, anno_col, ann_colors, order, scores_ordered)
}
\description{
Order samples and assemble heatmap inputs
}
